backend/requirements.txt
asgiref~=3.5.0
Django~=4.0.3
django-js-asset~=2.0.0
django-mathjax~=0.0.8
django-mptt~=0.13.4
django-mptt-admin~=2.2.0
psycopg2~=2.9.3
setuptools~=57.0.0
sqlparse~=0.4.2
tzdata~=2021.5
wheel~=0.36.2
djangorestframework~=3.14.0
rich~=12.6.0
djoser~=2.1.0
django-filter~=22.1.0
rest-framework-generic-relations~=2.1.0
django-cors-headers~=3.13.0
drf-flex-fields~=1.0.0